Refactor of app store rendering
