+++ src/reduxlike/reducer_remainder.js
@@ -8,8 +8,12 @@ import {
     ACTION_DO_NOTHING,
     ACTION_JOURNAL_ADD_ENTRY,
     ACTION_PHYSTYPE_DO_ACT,
+    ACTION_RENDER_QUEUE_ADD_GEO_CHART_DATA,
+    ACTION_RENDER_QUEUE_ADD_STATUS_MESSAGE,
+    ACTION_RENDER_QUEUE_ADD_TIME_CHART_DATA,
     ACTION_STORE_LOCK,
     ACTION_STORE_UNLOCK,
+    UI_NUM_TRAILS,
 } from '../const_vals.js';
 
 import {
@@ -19,7 +23,13 @@ import {
     simGetCurTime
 } from './store_getters.js';
 
-import { splice } from '../utils.js';
+import {
+    chartShiftData,
+    concatSliceMap,
+    fadeColors,
+    splice,
+} from '../utils.js';
+
 
 
 // *** Remainder reducer 
@@ -41,7 +51,7 @@ export const remainderReducer = (inStoreType) => (inActionType) =>
                 (getSavedPhysTypeStore(storeType))          // ... in this saved physType store...
                 (actionType.physType),                      // ... and replace with actionType.physType
         }),
-        
+
         [ACTION_DO_NOTHING]: (storeType) => (_) => ({ ...storeType.remainder }),
 
         [ACTION_JOURNAL_ADD_ENTRY]: (storeType) => (actionType) =>
@@ -84,6 +94,35 @@ export const remainderReducer = (inStoreType) => (inActionType) =>
             locked: false,
         }),
 
+        [ACTION_RENDER_QUEUE_ADD_STATUS_MESSAGE]: (storeType) => (actionType) =>
+        ({
+            ...storeType.remainder,
+            statusBoxContent: 'Time ' + roundTo(2)(simGetCurTime(storeType)) +
+                ': ' + actionType.msgStringType + '<br />',
+        }),
+
+        [ACTION_RENDER_QUEUE_ADD_GEO_CHART_DATA]: (storeType) => (actionType) =>
+        ({
+            ...storeType.remainder,
+            creature_geo_chart: mutable_updateGeoChartData(
+                storeType.remainder.creature_geo_chart,
+                actionType.dataIndexIntType,
+                actionType.colorStringType,
+                actionType.xyFloatTuple
+            ),
+        }),
+
+        [ACTION_RENDER_QUEUE_ADD_TIME_CHART_DATA]: (storeType) => (actionType) =>
+        ({
+            ...storeType.remainder,
+            creature_time_chart: mutable_updateTimeChartData(
+                storeType.remainder.creature_time_chart,
+                actionType.dataIndexIntType,
+                actionType.labelStringType,
+                actionType.timeValFloatTuple
+            ),
+        }),
+
         // use inActionType.type as an entry key into the key-val list above
         // key is used to select a function that takes a storeType object  
         //  and actionType and returns a storeType "remainder" prop obj
@@ -94,3 +133,91 @@ export const remainderReducer = (inStoreType) => (inActionType) =>
         //  and actionType to get a storeType "remainder" property object
         (inStoreType)
         (inActionType);
+
+
+// *** Reducer helper functions
+// update time history chart data
+// MUTABLE: mutates object at "chart" reference
+// takes: 
+//  chart: HTML DOM chart reference
+//  dataIndexIntType: chart data index, as int
+//  labelStringType: data label for legend, as string
+//  timeValFloatTuple: floating-point data point, as {time, value}
+// returns nothing
+export function mutable_updateTimeChartData(chart, dataIndexIntType, labelStringType, timeValFloatTuple) {
+    // MUTABLE: add data to chart
+    chart.data.datasets[dataIndexIntType] = {
+        ...chart.data.datasets[dataIndexIntType],
+        label: labelStringType,
+        data: chart.data.datasets[dataIndexIntType].data.concat(
+            {
+                x: timeValFloatTuple.time,
+                y: timeValFloatTuple.value
+            })
+    };
+
+    // revise time history chart x axis "window" if needed, for next chart update cycle
+    const chart_x = chart.options.scales.xAxes[0].ticks;                    // shorthand for x-axis ticks
+    const chart_xWidth = chart_x.max - chart_x.min;                         // extents of x axis
+    const new_max = Math.ceil(timeValFloatTuple.time - chart_x.stepSize);   // potential different x axis max           
+    const new_min = new_max - chart_xWidth;                                 // potential different x axis min
+
+    // MUTABLE: assign x axis min and max - shifted rightward if indicated by new_min and new_max
+    chart.options.scales.xAxes[0].ticks = {
+        ...chart_x,
+        max: (chart_x.max < new_max) ? new_max : chart_x.max,
+        min: (chart_x.min < new_min) ? new_min : chart_x.min,
+    };
+
+    // MUTABLE: shift out data that have "fallen off" the left side of the chart
+    chart.data.datasets[dataIndexIntType].data = chartShiftData
+        (new_min - chart.options.scales.xAxes[0].ticks.stepSize)
+        (chart.data.datasets[dataIndexIntType].data);
+
+    // return the passed-in chart reference
+    return chart;
+}
+
+// update geospatial chart data
+// MUTABLE: mutates object at "chart" reference
+// takes: 
+//  chart: HTML DOM chart reference
+//  dataIndexIntType: chart data index, as int
+//  colorStringType: color for data, as string
+//  xyFloatTuple: floating-point datapoint to add, as {x, y}
+// returns nothing
+function mutable_updateGeoChartData(chart, dataIndexIntType, colorStringType, xyFloatTuple) {
+    // all of our slice limits are -UI_NUM_TRAILS, so define a shorthand 
+    //  function with that limit built in 
+    const concatSliceTrailsMap = concatSliceMap(-UI_NUM_TRAILS);
+
+    // define a shorthand function specific to concatenating a color 
+    //  and mapping color list to a fade
+    const concatAndFade = concatSliceTrailsMap(fadeColors)(colorStringType);
+
+    // MUTABLE: add data and colors to chart, then slice to max length, 
+    //  then fade colors 
+    chart.data.datasets[dataIndexIntType] = {
+        ...chart.data.datasets[dataIndexIntType],
+
+        data: concatSliceTrailsMap
+            (x => x)                                            // identity function for mapping
+            ({ x: xyFloatTuple.x, y: xyFloatTuple.y })          // concatenate xyFloatTuple
+            ([chart.data.datasets[dataIndexIntType].data]),     // array: current chart xy data
+
+        backgroundColor:
+            concatAndFade([chart.data.datasets[dataIndexIntType].backgroundColor]),
+
+        borderColor:
+            concatAndFade([chart.data.datasets[dataIndexIntType].borderColor]),
+
+        pointBackgroundColor:
+            concatAndFade([chart.data.datasets[dataIndexIntType].pointBackgroundColor]),
+
+        pointBorderColor:
+            concatAndFade([chart.data.datasets[dataIndexIntType].pointBorderColor]),
+    };
+
+    // return the passed-in chart reference
+    return chart;
+}
