+++ src/reduxlike/store_init.js
@@ -13,13 +13,8 @@ import {
 
 import { actAsSimpleCreature } from '../creatures/simple_creature.js';
 import { mutableRandGen_initRandGen } from '../sim/seeded_rand.js';
-
-import { clearActionQueue, lockStore, unlockStore, actionDispatchPrivate } from './action_creators.js';
-import { applyArgChain } from '../utils.js';
-
-
-
-
+import { appStore } from './app_store.js';
+import { mutable_renderFunction } from './renderers.js';
 
 
 // *** Initial store
@@ -178,10 +173,10 @@ const initial_store = {
 
         // initial "saved physType" store
         savedPhysTypeStore: [{}],
-    },
 
-    // initial UI elements
-    ui: {
+        // initial status box content
+        statusBoxContent: '',
+
         // creature chart time reference placeholder
         creature_time_chart: null,
 
@@ -190,7 +185,7 @@ const initial_store = {
 
         // status box reference placeholder
         status_box: null,
-    }
+    },
 };
 
 
@@ -408,70 +403,6 @@ const creature_geo_chart_params_init = {
 };
 
 
-// *** Store constructor
-function Store(initialStore, creature_time_chart_context, creature_geo_chart_context, status_box_context) {
-    this.storeObj = {
-        ...initialStore,
-
-        // Simulator
-        sim: {
-            ...initialStore.sim,
-
-            // initRandGen just gives back the input seed
-            initSeed: mutableRandGen_initRandGen(initialStore.sim.initSeed),
-        },
-
-        // UI
-        ui: {
-            ...initialStore.ui,
-
-            // time chart
-            creature_time_chart: new Chart(creature_time_chart_context, creature_time_chart_params_init),
-
-            // geo chart
-            creature_geo_chart: new Chart(creature_geo_chart_context, creature_geo_chart_params_init),
-
-            // status box
-            status_box: status_box_context
-        }
-    };
-}
-
-
-
-// REFACTOR
-export const special_doActionQueue = (_) => ({
-    type: 'SPECIAL',
-});
-
-
-
-// *** Global app store
-export var appStore = {
-    storeObj: {},
-
-    // public action dispatch function
-    // takes:
-    //  storeType: app store, as storeType
-    //  ...actions: action creators to apply, each returning actionType
-    // returns storeType
-    actionDispatch: function (...actions) {
-        // chain up actions to dispatch
-        this.storeObj = applyArgChain
-            (actionDispatchPrivate)
-            (this.storeObj)
-            (
-                lockStore(),                // lock the store
-                actions,                    // dispatch given actions
-                special_doActionQueue(),    // dispatch actions in store queue
-                clearActionQueue(),         // clear action queue
-                unlockStore(),              // unlock the store
-            )
-    },
-}
-
-
-
 // *** Store initializer function
 export function storeInit(creature_time_chart_context, creature_geo_chart_context, status_box_context) {
     appStore.storeObj = {
@@ -486,8 +417,8 @@ export function storeInit(creature_time_chart_context, creature_geo_chart_contex
         },
 
         // UI
-        ui: {
-            ...initial_store.ui,
+        remainder: {
+            ...initial_store.remainder,
 
             // time chart
             creature_time_chart: new Chart(creature_time_chart_context, creature_time_chart_params_init),
@@ -496,7 +427,10 @@ export function storeInit(creature_time_chart_context, creature_geo_chart_contex
             creature_geo_chart: new Chart(creature_geo_chart_context, creature_geo_chart_params_init),
 
             // status box
-            status_box: status_box_context
+            status_box: status_box_context,
         }
     };
+
+    // set function to be called when the app store changes
+    appStore.setSubScribedFunc(mutable_renderFunction);
 }
